Compilation of other's branches

+++ Datalayer/DatabaseCalls.cs
@@ -15,55 +15,71 @@ public class DatabaseCalls : DBInterface
 
 
     //User Login/Registration things
-    public async Task<User> createUser(User user)
+    public async Task<User> createUser(string userName, string userPass, string email)
     {
-        if (await checkExisting(user))
+        
+        if (await checkExisting(userName))
         {
             //User Already Exists, throw an exception or something
             //THIS NEEDS TO BE DONE STILL
-            return user;
+            return new User();
         }
         else
         {
+            User newUser = new User()
+            {
+                username = userName,
+                password = userPass,
+                email = email
+            };
             //prob need to add other default values to user still
-            _context.Users.Add(user);
+            _context.Users.Add(newUser);
             await _context.SaveChangesAsync();
-            return user;
+            return newUser;
         }
 
     }
-    public async Task<User> loginUser(User auth)
+    public async Task<User> loginUser(string authUser, string authPass)
     {
-        if (await authenticateUser(auth))
+        if (await authenticateUser(authUser, authPass))
         {
             //If username and password match then give the user back
-            return await _context.Users.FirstOrDefaultAsync(user => user.username == auth.username && user.password == auth.password);
+            return await _context.Users.FirstOrDefaultAsync(user => user.username == authUser && user.password == authPass);
         }
         else
         {
             //if fail then throw an exception or something
-            return auth;
+            return new User();
         }
     }
-    public async Task<Boolean> checkExisting(User auth)
+    public async Task<Boolean> checkExisting(string authName)
     {
-        return await _context.Users.AnyAsync(user => user.username == auth.username);
+        return await _context.Users.AnyAsync(user => user.username == authName);
     }
-    public async Task<Boolean> authenticateUser(User auth)
+    public async Task<Boolean> authenticateUser(string authUser, string authPass)
     {
-        return await _context.Users.AnyAsync(user => user.username == auth.username && user.password == auth.password);
+        return await _context.Users.AnyAsync(user => user.username == authUser && user.password == authPass);
+    }
+
+    public async Task<User> updateUser(User auth)
+    {
+        User temp = await _context.Users.FirstOrDefaultAsync(user => user.username == auth.username);
+        Console.WriteLine(temp.bio);
+        temp.firstname = auth.firstname;
+        temp.lastname = auth.lastname;
+        temp.bio = auth.bio;
+        await _context.SaveChangesAsync();
+
+        Console.WriteLine(auth.bio);
+        return temp;
     }
 
     //************************************************ Post Related Things ************************************************
     public async Task<List<Post>> GetPostsByUserAsync(User user)
     {
-        List<Post> temp = await _context.Posts.FromSqlRaw($"SELECT * FROM Posts WHERE userId = {user.id}").ToListAsync();
-        List<Post> userx = new List<Post>();
-        foreach (Post p in temp)
-        {
-            userx.Add(await _context.Posts.FirstOrDefaultAsync(user => user.id == p.id));
-        }
-        return userx;
+        // List<Post> temp = await _context.Posts.FromSqlRaw($"SELECT * FROM Posts WHERE userId = {user.id}").ToListAsync();
+        // return temp;
+        return await _context.Posts.FromSqlRaw($"SELECT * FROM Posts WHERE userId = {user.id}").ToListAsync();
     }
     public async Task<List<Post>> getPostbyUserIdAsync(int userId)
     {
@@ -75,22 +91,22 @@ public class DatabaseCalls : DBInterface
         return await _context.Posts.AsNoTracking().Where(post => post.bandId == bandId).ToListAsync();
     }
 
-    public async Task postForBandAsync(int bandId, string textEntry)
+    public async Task postForBandAsync(int bandId, string textEntry, string postType)
     {
-        Post post = new Post() { entry = textEntry, bandId = bandId };
+        Post post = new Post() { entry = textEntry, bandId = bandId, type = postType };
         _context.Posts.Add(post);
         await _context.SaveChangesAsync();
     }
 
-    public async Task postForUserAsync(User user, string textEntry)
+    public async Task postForUserAsync(User user, string textEntry, string postType)
     {
-        Post post = new Post() { entry = textEntry, userId = user.id, likes = 0, dateCreated = DateTime.Now };
+        Post post = new Post() { entry = textEntry, userId = user.id, likes = 0, dateCreated = DateTime.Now, type = postType };
         _context.Posts.Add(post);
         await _context.SaveChangesAsync();
     }
-    public async Task postForUserIdAsync(int userId, string textEntry)
+    public async Task postForUserIdAsync(int userId, string textEntry, string postType)
     {
-        Post post = new Post() { entry = textEntry, userId = userId, likes = 0, dateCreated = DateTime.Now };
+        Post post = new Post() { entry = textEntry, userId = userId, likes = 0, dateCreated = DateTime.Now, type = postType };
         _context.Posts.Add(post);
         await _context.SaveChangesAsync();
     }
@@ -99,8 +115,11 @@ public class DatabaseCalls : DBInterface
     {
         // might not work, just put it here for now. Also allows for inifite likes
         Post temp = await _context.Posts.FirstOrDefaultAsync(t => t.id == postId);
-        temp.likes++;
-        _context.Posts.Update(temp);
+        if(temp != null)
+        {
+            temp.likes++;
+            _context.Posts.Update(temp);
+        }
         //_context.Posts.FromSqlRaw($"UPDATE Posts SET likes = likes + 1 WHERE id = {postId}");
         await _context.SaveChangesAsync();
     }
@@ -108,12 +127,11 @@ public class DatabaseCalls : DBInterface
     public async Task deletePostAsync(int postId)
     {
         //_context.Posts.FromSqlRaw($"DELETE from Posts WHERE id={postId}");
-        Post temp = await _context.Posts.FirstOrDefaultAsync(t => t.id == postId);
-        /*if(temp == null)
+        Post temp = await _context.Posts.FirstOrDefaultAsync(t=> t.id == postId);
+        if(temp != null)
         {
-            return Task.CompletedTask;
-        }*/
         _context.Posts.Remove(temp);
+        }
         await _context.SaveChangesAsync();
     }
 
@@ -165,11 +183,32 @@ public class DatabaseCalls : DBInterface
 
     //Group/Team things
 
+    //Band calls -Arrion
+    public async Task<Band> CreateBand(Band newBand)
+    {
+        _context.Bands.Add(newBand);
+        await _context.SaveChangesAsync();
+        return newBand;
+    }
+
+    public async Task<List<Band>> GetBands(int bandId)
+    {
+        return await _context.Bands.FromSqlRaw($"Select * From Band Where Bands.bandId = {bandId}").ToListAsync();
+    }
+
+    public async Task DeleteBand(Band bandToDelete)
+    {
+        _context.Bands.Remove(bandToDelete);
+        await _context.SaveChangesAsync();
+    }
+
     // Gets all members for a single Band ~Bailey
     public async Task<List<BandMember>> GetAllBandMembers(int bandId)
     {
         return await _context.BandMembers.FromSqlRaw($"Select * From BandMembers Where BandMembers.bandId = {bandId}").ToListAsync();
     }
+
+    //Note to me(Arrion) don't forget band limit
 
     // Adds new member to database ~Bailey
     public async Task<BandMember> CreateBandMember(BandMember newMember)
@@ -184,6 +223,16 @@ public class DatabaseCalls : DBInterface
     {
         _context.BandMembers.Remove(memberToDelete);
         await _context.SaveChangesAsync();
+    }
+
+    public Task<bool> authenticateUser(User user)
+    {
+        throw new NotImplementedException();
+    }
+
+    public Task<bool> checkExisting(User user)
+    {
+        throw new NotImplementedException();
     }
 }
